requirements.txt
streamlit>=1.32.0
streamlit-authenticator>=0.4.0
bcrypt>=4.1.2
pyyaml>=6.0.1
plotly>=5.18.0
psycopg2-binary>=2.9.9
pymssql>=2.2.11
pandas>=2.1.0
openpyxl>=3.1.2
